LPS-151680 Update tree tests
