+++ modules/apps/content-dashboard/content-dashboard-web/test/js/components/SelectFileExtension.test.js
@@ -91,7 +91,7 @@ describe('SelectFileExtension', () => {
 		);
 
 		const {className} = getByRole('tree');
-		expect(className).toContain('lfr-treeview-node-list');
+		expect(className).toContain('treeview treeview-light');
 
 		expect(
 			getByText('Audio (5 items)', {exact: false})
