+++ modules/apps/content-dashboard/content-dashboard-web/test/js/components/SelectTypeAndSubtype.test.js
@@ -84,7 +84,7 @@ describe('SelectTypeAndSubtype', () => {
 		);
 
 		const {className} = getByRole('tree');
-		expect(className).toContain('lfr-treeview-node-list');
+		expect(className).toContain('treeview treeview-light');
 
 		expect(
 			getByText('Document (3 items)', {exact: false})
